tools/nxfuse: Fix svace errors
- Fix - possibly uninitialized variable 'nxfuse_data' passed to fuse_main()
- Fix - check possible NULL return value of recall before deferencing pointer
- Fix - file opened but not closed before returning in error case
- Fix - assigning value of signed int to bigger unsigned int
- Fix - use 'strncpy' instead of 'strcpy' for safety
- Fix - possibly unreachable code

+++ tools/nxfuse/src/nxfuse.h
@@ -109,6 +109,6 @@ struct inode *vmount(const char *filename, const char *mount_point, const char *
  *   specified NuttX filesystem type.
  *
  ****************************************************************************/
-int mkfs(const char *filename, const char *fs_type, int erasesize, int sectsize, int pagesize, char *generic, int confirm);
+int mkfs(const char *filename, const char *fs_type, uint32_t erasesize, uint16_t sectsize, int pagesize, char *generic, int confirm);
 
 #endif							/* _SRC_NXFUSE_H */
